# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tier Trade API
  version: 0.1.0
security:
  - bearerAuth: []
paths:
...
  /v1/inventory/sales-contracts:
    post:
      summary: Cria um contrato de venda ativo
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SalesContractInput' } } }
      responses:
        '201': { description: Contrato de venda criado }
        '409': { description: Referência já utilizada }
  /v1/inventory/sales-contracts/{contractId}:
    put:
      summary: Atualiza contrato operacional de venda, preservando a versão anterior; não formaliza aditivo assinado
      parameters:
        - { name: contractId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SalesContractInput' } } }
      responses:
        '200': { description: Contrato atualizado e nova versão registrada }
        '422': { description: Volume inferior ao saldo já alocado }
  /v1/inventory/sales-contracts/{contractId}/versions:
    get:
      summary: Lista snapshots imutáveis dos termos operacionais, do mais recente ao mais antigo
      parameters:
        - { name: contractId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Versões do contrato acessíveis somente ao tenant autenticado }
        '404': { description: Contrato inexistente no tenant }
  /v1/inventory/allocations:
    post:
      summary: Reserva saldo disponível de um lote para contrato de venda
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AllocationInput' } } }
      responses:
        '201': { description: Alocação criada }
        '422': { description: Saldo de lote ou contrato insuficiente }
  /v1/inventory/allocations/{allocationId}/release:
    post:
      summary: Libera alocação ativa que ainda não possui expedições
      parameters:
        - { name: allocationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Alocação liberada }
        '409': { description: Alocação já executada ou inativa }
  /v1/inventory/dispatches:
    post:
      summary: Registra expedição e movimento negativo de estoque atomicamente
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DispatchInput' } } }
      responses:
        '201': { description: Expedição confirmada e previsão financeira criada atomicamente }
        '422': { description: Saldo alocado ou físico insuficiente }
...
  /v1/inventory/locations:
    post:
      summary: Cadastra local físico de estoque do tenant
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateInventoryLocation' } } }
      responses: { '201': { description: Local cadastrado } }
  /v1/inventory/lots/{lotId}/classification:
    put:
      summary: Classifica propriedade, risco e custódia do lote
      parameters: [{ $ref: '#/components/parameters/LotId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ClassifyInventoryLot' } } }
      responses: { '200': { description: Classificação atualizada e auditada } }
  /v1/inventory/lots/{lotId}/transfers:
    post:
      summary: Inicia remaneio físico do lote
      parameters: [{ $ref: '#/components/parameters/LotId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/StartInventoryTransfer' } } }
      responses: { '201': { description: Remaneio iniciado } }
  /v1/inventory/transfers/{transferId}/complete:
    post:
      summary: Conclui remaneio e atualiza o local do lote
      parameters:
        - { name: transferId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CompleteInventoryTransfer' } } }
      responses: { '201': { description: Remaneio concluído } }
  /v1/inventory/lots/{lotId}/losses:
    post:
      summary: Registra perda física sem permitir saldo negativo ou quebrar alocações
      parameters: [{ $ref: '#/components/parameters/LotId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/InventoryLoss' } } }
      responses:
        '201': { description: Perda registrada }
        '409': { description: Perda incompatível com o saldo ou as alocações ativas }
  /v1/inventory/lots/{lotId}/counts:
    post:
      summary: Reconcilia contagem física com o razão de estoque
      parameters: [{ $ref: '#/components/parameters/LotId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/InventoryCount' } } }
      responses:
        '201': { description: Contagem e eventual ajuste registrados }
        '409': { description: Contagem incompatível com alocações ativas }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    LotId:
      name: lotId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    CreateInventoryLocation:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{1,31}$' }
        name: { type: string, minLength: 2, maxLength: 120 }
    ClassifyInventoryLot:
      type: object
      additionalProperties: false
      required: [ownershipStatus, riskStatus, custodyStatus, ownerCounterpartyId, custodianCounterpartyId, reason, occurredAt]
      properties:
        ownershipStatus: { enum: [PENDING_DEFINITION, OWN, THIRD_PARTY] }
        riskStatus: { enum: [PENDING_DEFINITION, ASSUMED, NOT_ASSUMED] }
        custodyStatus: { enum: [IN_STORAGE, IN_TRANSIT, RELEASED] }
        ownerCounterpartyId: { type: [string, 'null'], format: uuid }
        custodianCounterpartyId: { type: [string, 'null'], format: uuid }
        reason: { type: string, minLength: 5, maxLength: 1000 }
        occurredAt: { type: string, format: date-time }
    StartInventoryTransfer:
      type: object
      additionalProperties: false
      required: [destinationLocationId, reason, startedAt]
      properties:
        destinationLocationId: { type: string, format: uuid }
        reason: { type: string, minLength: 5, maxLength: 1000 }
        startedAt: { type: string, format: date-time }
    CompleteInventoryTransfer:
      type: object
      additionalProperties: false
      required: [completedAt, reason]
      properties:
        completedAt: { type: string, format: date-time }
        reason: { type: string, minLength: 5, maxLength: 1000 }
    InventoryLoss:
      type: object
      additionalProperties: false
      required: [quantityKg, occurredAt, reason]
      properties:
        quantityKg: { type: string, pattern: '^\d+(?:\.\d{1,3})?$' }
        occurredAt: { type: string, format: date-time }
        reason: { type: string, minLength: 5, maxLength: 1000 }
    InventoryCount:
      type: object
      additionalProperties: false
      required: [countedQuantityKg, occurredAt, reason]
      properties:
        countedQuantityKg: { type: string, pattern: '^\d+(?:\.\d{1,3})?$' }
        occurredAt: { type: string, format: date-time }
        reason: { type: string, minLength: 5, maxLength: 1000 }
...
    DecimalString: { type: string, pattern: '^\d+(?:\.\d{1,6})?$' }
...
    SalesContractInput:
      type: object
      additionalProperties: false
      required: [counterpartyId, reference, commodity, quantityKg, salePricePerKg, destinationCode, deliveryStart, deliveryEnd, requiredDocuments]
      properties:
        counterpartyId: { type: string, format: uuid }
        reference: { type: string, minLength: 3, maxLength: 40 }
        commodity: { enum: [MILHO, SOJA] }
        quantityKg: { $ref: '#/components/schemas/DecimalString' }
        salePricePerKg: { $ref: '#/components/schemas/DecimalString' }
        destinationCode: { type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{1,31}$' }
        deliveryStart: { type: string, format: date }
        deliveryEnd: { type: string, format: date }
        requiredDocuments: { type: array, maxItems: 20, items: { type: string, maxLength: 80 } }
        paymentTermDays: { type: [integer, 'null'], minimum: 0, maximum: 730 }
    AllocationInput:
      type: object
      additionalProperties: false
      required: [salesContractId, lotId, quantityKg]
      properties:
        salesContractId: { type: string, format: uuid }
        lotId: { type: string, format: uuid }
        quantityKg: { $ref: '#/components/schemas/DecimalString' }
    DispatchInput:
      type: object
      additionalProperties: false
      required: [allocationId, quantityKg, dispatchedAt, vehiclePlate, documentReference, notes]
      properties:
        allocationId: { type: string, format: uuid }
        quantityKg: { $ref: '#/components/schemas/DecimalString' }
        dispatchedAt: { type: string, format: date-time }
        vehiclePlate: { type: string, pattern: '^[A-Z]{3}[0-9][A-Z0-9][0-9]{2}$' }
        documentReference: { type: string, minLength: 1, maxLength: 80 }
        notes: { type: [string, 'null'], maxLength: 1000 }
...